tasks added
//выдавать экранные координаты
//выдавать значения сплайна
//нормальный зум

+++ graphics2/untitled1/glview.hpp
@@ -8,8 +8,7 @@
 #include <vector>
 #include <cmath>
 #include <iostream>
-
-float CubicHermite (float A, float B, float C, float D, float x);
+double CubicHermite (float A, float B, float C, float D, float x);
 
 class glView : public QGLWidget
 {
